feat : Choice엔티티에 voteCount 추가한다.

+++ src/main/java/life/offonoff/ab/web/response/topic/choice/ChoiceResponseWithCount.java
@@ -9,5 +9,5 @@ import lombok.Getter;
 public class ChoiceCountResponse {
 
     private ChoiceOption choiceOption;
-    private Long voteCount;
+    private int voteCount;
 }
